Share affordable fee calculation
